+++ common/src/main/java/com/slymask3/instantblocks/block/instant/InstantLavaBlock.java
@@ -21,10 +21,9 @@ public class InstantLavaBlock extends InstantLiquidBlock {
 				.sound(new LiquidSoundType(SoundEvents.BUCKET_FILL_LAVA, SoundEvents.BUCKET_EMPTY_LAVA))
 				.lightLevel((par1) -> 15)
 		, Blocks.AIR, Blocks.LAVA);
-		setErrorMessage(Strings.ERROR_LAVA_MAX);
-		this.create = Strings.CREATE_LAVA;
-		this.create1 = Strings.CREATE_LAVA_1;
-		this.particle = ParticleTypes.LARGE_SMOKE;
+		this.setErrorMessage(Strings.ERROR_LAVA_MAX);
+		this.setCreateMessages(Strings.CREATE_LAVA,Strings.CREATE_LAVA_1);
+		this.setParticle(ParticleTypes.LARGE_SMOKE);
 	}
 
 	public boolean isEnabled() {